Kits almost complete

+++ database/migrations/2018_05_21_144944_create_taqmanprocurements_table.php
@@ -18,7 +18,7 @@ class CreateTaqmanprocurementsTable extends Migration
         Schema::create('taqmanprocurements', function (Blueprint $table) {
             $table->increments('id');
             $table->tinyInteger('month')->unsigned()->index();
-            $table->tinyInteger('year')->unsigned()->index();
+            $table->integer('year')->unsigned()->index();
             $table->tinyInteger('testtype')->unsigned()->index();
             $table->integer('received')->unsigned()->default(0)->nullable();
             $table->integer('tests')->unsigned()->default(0)->nullable();
