+++ database/migrations/2018_05_22_143158_create_abbotdeliveries_table.php
@@ -20,7 +20,7 @@ class CreateAbbotdeliveriesTable extends Migration
             $table->tinyInteger('testtype')->unsigned()->index();
             $table->tinyInteger('lab')->unsigned()->index();
             $table->tinyInteger('quarter')->unsigned()->index();
-            $table->tinyInteger('year')->unsigned()->index();
+            $table->integer('year')->unsigned()->index();
             $table->tinyInteger('source')->unsigned()->nullable();
             $table->tinyInteger('labfrom')->unsigned()->nullable();
             foreach ($this->main as $key => $value) {
